Extract the validation of Jira integrations to prepare for team reuse
