+++ server/service/service_appconfig.go
@@ -82,25 +82,6 @@ func (svc *Service) sendTestEmail(ctx context.Context, config *fleet.AppConfig) 
 	return nil
 }
 
-func (svc *Service) makeTestJiraRequest(ctx context.Context, jiraSettings *fleet.JiraIntegration) error {
-	if jiraSettings.APIToken == "" || jiraSettings.APIToken == fleet.MaskedPassword {
-		return &badRequestError{message: fmt.Sprintf("jira integration request failed: missing or invalid API token")}
-	}
-	client, err := externalsvc.NewJiraClient(&externalsvc.JiraOptions{
-		BaseURL:           jiraSettings.URL,
-		BasicAuthUsername: jiraSettings.Username,
-		BasicAuthPassword: jiraSettings.APIToken,
-		ProjectKey:        jiraSettings.ProjectKey,
-	})
-	if err != nil {
-		return &badRequestError{message: fmt.Sprintf("jira integration request failed: %s", err.Error())}
-	}
-	if _, err := client.GetProject(ctx); err != nil {
-		return &badRequestError{message: fmt.Sprintf("jira integration request failed: %s", err.Error())}
-	}
-	return nil
-}
-
 func (svc *Service) makeTestZendeskRequest(ctx context.Context, zendeskSettings *fleet.ZendeskIntegration) error {
 	if zendeskSettings.APIToken == "" || zendeskSettings.APIToken == fleet.MaskedPassword {
 		return &badRequestError{message: fmt.Sprintf("zendesk integration request failed: missing or invalid API token")}
